-Now works better if one of your pages uses autolayout.
Previously if one of your pages used autolayout (or had a MKMapView or UIScrollView on it which also seemed to use some autolayout bits), the zoom animation as you scrolled between pages wouldn’t work and you would be stuck on the same page. This was because the transform property on UIView doesn’t work well if the view uses autolayout. Now, before a page is added to the scrollview in this control, it is embedded inside another uiview which shields it from this problem. This should be transparent to you and not require any changes to people using the control.

Fixes #4